file service in progress

+++ srcs/backend/main_service/src/services/userService.js
@@ -2,10 +2,9 @@ const userRepository = require('../repositories/userRepository')
 const argon2 = require('argon2');
 const {HttpException} = require('../utils/httpExceptions');
 const crypto = require('crypto');
+const fileService =  require('./fileService');
 
-
-const createUser = async (userData) =>
-{
+const createUser = async (userData) => {
     const {password, ...data} = userData;
     const passwordHash = await argon2.hash(password);
     const user = await userRepository.createUser({passwordHash, ...data})
@@ -13,8 +12,7 @@ const createUser = async (userData) =>
     return user;
 }
 
-const findUserOrCreate = async (profile) =>
-{
+const findUserOrCreate = async (profile) => {
     const email = profile.emails[0].value. toLowerCase().trim();
     let user = await userRepository.getByEmail(email);
     if (user) {
@@ -39,24 +37,21 @@ const findUserOrCreate = async (profile) =>
     return user;
 }
 
-const getUserById = async (userId) =>
-{
+const getUserById = async (userId) => {
     const user = await userRepository.getUserById(userId);
     if (!user)
         return;
     return user;
 }
 
-const getUserByEmail = async (email) =>
-{
+const getUserByEmail = async (email) => {
     const user = await userRepository.getByEmail(email);
     if (!user)
         return;
     return user;
 }
 
-const updateUser = async (userId, updateData) =>
-{
+const updateUser = async (userId, updateData) => {
     await getUserById(userId);
     const allowedFields = ['firstName', 'lastName', 'phone', 'avatarUrl','refreshToken', "isVerified"];
     const filteredData = {};
@@ -70,18 +65,45 @@ const updateUser = async (userId, updateData) =>
     return await userRepository.updateUser(userId,filteredData);
 }
 
-const deleteUser = async (userId) =>
-{
-    await getUserById(userId);
-    //handle failure
+const deleteUser = async (userId) => {
+    const user =  await getUserById(userId);
+    if (!user)
+        throw new HttpException(404, "user not found");
     await userRepository.deleteUser(userId);
 }
 
-const getUsers = async (filters) =>
-{
+const getUsers = async (filters) => {
     return await userRepository.getUsers(filters);
 }
 
+const uploadAvatar = async (userId, file) => {
+    const user = await userRepository.getUserById(userId);
+    const {avatarUrl} =  await fileService.saveAvatar(userId,file);
+    if (avatarUrl !== user.avatarUrl) {
+        await fileService.deleteFile(user.avatarUrl);
+    }
+    const updatedUser =  userRepository.updateUser(userId, {avatarUrl});
+    return updateUser;
+
+}
+const detletAvatar =  async (userId) => {
+    const user = userRepository.getUserById(userId);
+    if (!user)
+        throw new HttpException(404, "user not found");
+    if (!user.avatarUrl)
+        throw new HttpException(400, "avatar not set yet");
+    await fileService.deleteFile(user.avatarUrl);
+    await userRepository.updateUser(userId, {avatarUrl : null});
+}
+
+const getAvatar = async (userId) => {
+    const user =  userRepository.getUserById(userId);
+    if (!user)
+        throw new HttpException(404, 'user not found');
+    if (!user.avatarUrl)
+        throw new HttpException(404, 'avatar not setted yet');
+    return user.avatarUrl;
+}
 module.exports = {
     createUser,
     findUserOrCreate,
@@ -89,5 +111,8 @@ module.exports = {
     getUsers,
     updateUser,
     getUserByEmail,
-    getUserById
+    getUserById,
+    getAvatar,
+    detletAvatar,
+    uploadAvatar,
 };
